requirements.txt
streamlit
numpy
pandas
plotly
requests
yfinance
fpdf
scikit-learn
python-dotenv
